tesseract-ocr-app/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-CORS==4.0.0
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
Pillow>=10.0.0
pytesseract>=0.3.10
Werkzeug==3.1.3
numpy<2.0.0
opencv-python>=4.8.0,<4.9.0
reportlab>=4.0.0
PyPDF2>=3.0.0
